chore(coin): partially move back cache.ts into live-common
