+++ libs/ledger-live-common/package.json
@@ -208,6 +208,7 @@
     "jwt-decode": "^3.1.2",
     "leb128": "^0.0.5",
     "lodash": "^4.17.21",
+    "lru-cache": "^7.14.1",
     "minimatch": "^5.1.0",
     "near-api-js": "^0.44.2",
     "numeral": "^2.0.6",
