+++ libs/ledger-live-common/src/families/polkadot/bridge/js.ts
@@ -4,8 +4,9 @@ import {
 } from "@ledgerhq/coin-polkadot/bridge/js";
 import { withDevice } from "../../../hw/deviceAccess";
 import network from "../../../network";
+import { makeLRUCache } from "../../../cache";
 
 export default {
-  currencyBridge: buildCurrencyBridge(withDevice, network),
-  accountBridge: buildAccountBridge(withDevice, network),
+  currencyBridge: buildCurrencyBridge(withDevice, network, makeLRUCache),
+  accountBridge: buildAccountBridge(withDevice, network, makeLRUCache),
 };
